Moved dtos in dtos module.
