+++ dtos/src/main/java/org/eclipse/dataspaceconnector/identityhub/dtos/MessageResponseObject.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.identityhub.models;
+package org.eclipse.dataspaceconnector.identityhub.dtos;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
