+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/api/IdentityHubController.java
@@ -20,17 +20,17 @@ import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
-import org.eclipse.dataspaceconnector.identityhub.models.MessageRequestObject;
-import org.eclipse.dataspaceconnector.identityhub.models.MessageResponseObject;
-import org.eclipse.dataspaceconnector.identityhub.models.RequestObject;
-import org.eclipse.dataspaceconnector.identityhub.models.RequestStatus;
-import org.eclipse.dataspaceconnector.identityhub.models.ResponseObject;
-import org.eclipse.dataspaceconnector.identityhub.models.WebNodeInterfaces;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageRequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.RequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.RequestStatus;
+import org.eclipse.dataspaceconnector.identityhub.dtos.ResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.WebNodeInterfaces;
 import org.eclipse.dataspaceconnector.identityhub.processor.MessageProcessorRegistry;
 
 import java.util.stream.Collectors;
 
-import static org.eclipse.dataspaceconnector.identityhub.models.WebNodeInterfaceMethod.fromName;
+import static org.eclipse.dataspaceconnector.identityhub.dtos.WebNodeInterfaceMethod.fromName;
 
 /**
  * Identity Hub controller, exposing a <a href="https://identity.foundation/decentralized-web-node/spec">Decentralized Web Node</a> compatible endpoint.
