+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/processor/InterfaceNotImplementedProcessor.java
@@ -14,12 +14,12 @@
 
 package org.eclipse.dataspaceconnector.identityhub.processor;
 
-import org.eclipse.dataspaceconnector.identityhub.models.MessageResponseObject;
-import org.eclipse.dataspaceconnector.identityhub.models.MessageStatus;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageStatus;
 
 import java.util.List;
 
-import static org.eclipse.dataspaceconnector.identityhub.models.MessageResponseObject.MESSAGE_ID_VALUE;
+import static org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject.MESSAGE_ID_VALUE;
 
 /**
  * Default message processor when a non-supported interface is provided
