+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/processor/MessageProcessor.java
@@ -14,8 +14,8 @@
 
 package org.eclipse.dataspaceconnector.identityhub.processor;
 
-import org.eclipse.dataspaceconnector.identityhub.models.MessageResponseObject;
-import org.eclipse.dataspaceconnector.identityhub.models.WebNodeInterfaces;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.WebNodeInterfaces;
 
 /**
  * <p>
