+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include(":spi:identity-hub-spi")
 include(":spi:identity-hub-store-spi")
 include(":extensions:identity-hub")
 include(":identity-hub-client")
+include("dtos")
